+++ angular-shortcuts.js
@@ -8,7 +8,8 @@
   function $shortCuts(){
     var provider = {
       $get : $get,
-      shortCut : shortCut
+      shortCut : shortCut,
+      showLog : showLog
     }
     , _keyCodes = {
       'a' : 65,
@@ -38,7 +39,8 @@
       'y' : 89,
       'z' : 90,
     }
-    , _shortCuts = [];
+    , _shortCuts = []
+    , _showLog = false;
     return provider;
 
     function shortCut(keys, config) {
@@ -46,13 +48,19 @@
       return this;
     };
 
+    function showLog(option){
+      _showLog = option;
+    };
+
     function ShortCut(keys, config){
       var keys = buildKeys(keys);
-      validateKeys(keys);
+      validateKeysAmount(keys);
+      validateConfig(config);
 
       var shortCut = {
           keyOne : getKeyCode(keys[0])
         , keyTwo : getKeyCode(keys[1])
+        , config : config
       };
 
       return shortCut;
@@ -70,12 +78,30 @@
       return keys.split('+');
     };
 
-    function validateKeys(keys){
-      if (keys.length > 2)
-        throw new InvalidKeysException();
+    function log(mgs){
+      if (_showLog) {
+        console.warn(mgs);
+      };
     };
 
-    function InvalidKeysException() {
+    function validateConfig(config){
+      if (!config.actionType){
+        throw new ActionTypeNotDefined();
+      };
+      if (config.actionType === 'EVENT' && !config.eventName){
+        throw new EventNameNotDefined();
+      };
+      if (config.actionType === 'CALLBACK' && !angular.isFunction(config.callBack)){
+        throw new CallBackNotDefined();
+      };
+    };
+
+    function validateKeysAmount(keys){
+      if (keys.length > 2)
+        throw new InvalidKeysAmountException();
+    };
+
+    function InvalidKeysAmountException() {
       return Error('More than two keys defined.');
     };
   
@@ -83,20 +109,34 @@
       return Error('The key "' + key + '" is not available.');
     };
 
+    function ActionTypeNotDefined(){
+      return Error('Action type not defined in config.');
+    };
+
+    function EventNameNotDefined(){
+      return Error('Event name not defined in config.');
+    };
+
+    function CallBackNotDefined(){
+      return Error('CallBack not defined or not is a function.')
+    };
+
     // the $get function retuns the service
     $get.$inject = ['$log', '$rootScope'];
-    function $get($log, $rootScope, $document){
+    function $get($log, $rootScope){
       var service = {
-        log : log
+        getShortCuts : getShortCuts
       }
       , fisrtKeyActive = false
       , firstKeyCode;
-
       document.onkeydown = onKeyDown; 
 
+      return service;
+
       function onKeyDown(event){
-        if (event.target.nodeName === 'BODY')
+        if (event.target.nodeName === 'BODY') {
           handle(event);
+        };
       };
 
       function handle(event){
@@ -106,25 +146,50 @@
         };
         
         if (fisrtKeyActive == true) {
-
+          var shortCut = geShortCutByKeys(firstKeyCode, event.keyCode);
+          if (shortCut) {
+            executeConfigAction(shortCut.config);
+            fisrtKeyActive = false;
+          };
         };
+      };
+
+      function executeConfigAction(config){
+        if (config.actionType === 'EVENT'){
+          dispachEvent(config.eventName);
+          return;
+        };
+        log('Running callBack {} ' + config.callBack);
+        config.callBack();
+      };
+
+      function dispachEvent(eventName){
+        log('Broadcasting event {} "' + eventName + '".')
+        $rootScope.$broadcast(eventName);
       };
 
       function isFirstKey(key){
         var result = false;
         angular.forEach(_shortCuts, function(shortCut) {
-          if (shortCut[keyType] === key)
+          if (shortCut.keyOne === key)
             result = true;
         });
         return result;
       };
 
-      // for test
-      function log(){
-        $log.info('wow', _shortCuts);
+      function geShortCutByKeys(keyOne, keyTwo){
+        var result;
+        angular.forEach(_shortCuts, function(shortCut) {
+          if (shortCut.keyOne === keyOne && shortCut.keyTwo === keyTwo)
+            result = shortCut;
+        });
+        return result;
+      };
+
+      function getShortCuts(){
+        return angular.copy(_shortCuts);
       };
       
-      return service;
     };
   };
 })();