+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## [0.0.3] - 2015-07-08
+### Added
+- Validations on create the shortcut object and add news exceptions
+- Broadcast defined event on the execution of shortcut
+- Running defined callBack on the execution of shortcut
+- Method showLog to define the log option
+
 ## [0.0.2] - 2015-07-04
 ### Changed
 - Starting the validations in the provider.